+++ src/constants/regex.ts
@@ -4,3 +4,6 @@ export const email: RegExp = /^[^\s@]+@[^\s@]+\.([^\s@]{2,})+$/;
 
 export const password: RegExp =
   /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[0-9a-zA-Z]{6,}$/;
+
+
+export const notification: RegExp = /[a-zA-Z\ ]{6,15}/;