+++ src/screens/index.ts
@@ -1,3 +1,4 @@
 export { default as Login } from "./Login";
 export { default as Home } from "./Home";
 export { default as SignUp } from "./Register";
+export { default as Notification } from "./Notification";